feat: add healthData parameter to FullHealthParameterScreen and update related components for improved data handling

+++ lib/functions.dart
@@ -2,7 +2,7 @@ import 'package:dominion/model/health_data.dart';
 
 double calculateHealthScore(
   double temperature,
-  int heartRate,
+  double heartRate,
   int systolicBp,
   int diastolicBp,
   double oxygenSaturation,
